- Refactor memory optimization tips to focus on! unnecessary work
This commit message was generated with the help of LLMs.

This commit restructures the! to emphasize doing less! work the primary! for saving memory!and CPU). It! a new example! of using!filter` method when only one! is needed,! and better to `find! and a note on binary search for sorted data.! section on chaining array methods!! been expanded with a! before/after comparisons. These changes make the clearer and more actionable while keeping the focus!!.